cxl/acpi: cfmws: fix Base HBA and Window Size
and update print format.

in hostos: $ ./vm.sh

then, in guestos:

    rongtao@fedora-vm-cxl:~/Git/tst-linux/cxl/acpi$ sudo ./cedt
    ---------------- CEDT ----------------
    CEDT length 108
    Signature:	CEDT
    Revision:	0x1
    Checksum:	0x6f
    OEM ID:	'BOCHS '(raw:424f43485320)
    OEM Table ID:	'BXPC    '(raw:4258504320202020)
    OEM Revision:	0x1
    Creator ID:	0x43505842
    Creator Revision:	0x1
    CEDT structure size:	72
    struct cedt header size:	36
    ---------------- CHBS ----------------
    struct chbs size 32
    CHBS type 0, len 32, uid b, CXL version 1, base a80000000 len 10000
    ---------------- CFMWS ----------------
    struct cfmws size 40
    CFMWS type 1, record length 40(0x28), hpa 0xa90000000, winsize 4294967296(4096 MiB, 4 GiB), ENIW 0, hbig 4 (0 MB), qtag id 0, NIW 1, ...

Envs: Arch x86_64, kvm
Vers: Fedora Linux 44 (Forty Four), Linux 6.14.11-300.fc42_cxl.x86_64, acpica-tools 20260408, ACPITool 0.5.1, GCC 16.1.1, GLibc 2.43, NDCtl 85, Python 3.14.6

+++ cxl/acpi/cfmws.h
@@ -2,7 +2,10 @@
 // Copyright (C) 2025-2026 Rong Tao
 /**
  * CXL Fixed Memory Window Structure (CFMWS)
- * CXL 3.0 Specification, Section 9.17.1.3
+ * CXL 3.0 Specification, Section 9.17.1.3, Table 9-22.
+ *
+ * Abbrev:
+ * - NIW: Number of Interleave Ways
  */
 #pragma once
 #include <stdint.h>
@@ -15,10 +18,15 @@ struct cfmws {
 	 */
 	uint16_t record_length;
 	uint32_t reserved2;
-	uint8_t base_hpa;
-	uint8_t window_size;
+	uint64_t base_hpa;
+	uint64_t window_size;
 	/**
 	 * Encoded Number of Interleave Ways (ENIW)
+	 *
+	 * NIW is the raw count of Interleave ways whereas ENIW is the encoded
+	 * value:
+	 * - If ENIW <  8, NIW = 2^ENIW
+	 * - If ENIW >= 8, NIW = 3 * 2^(ENIW - 8)
 	 */
 	uint8_t eniw;
 	uint8_t interleave_arithmetic;
@@ -35,7 +43,8 @@ struct cfmws {
 	/* QTAG ID */
 	uint16_t qtag_id;
 	/**
-	 * Number of Interleave Ways (NIW)
+	 * A list of all the Interleave Targets.
+	 * size = 4 * NIW
 	 */
 	uint32_t interleave_target_list[];
 };
